% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaAnalysisVariant-methods.R
\docType{methods}
\name{OpencgaAnalysisVariant,Opencga-method}
\alias{OpencgaAnalysisVariant,Opencga-method}
\alias{OpencgaUser}
\title{A method to query user data}
\usage{
\S4method{OpencgaAnalysisVariant}{Opencga}(object, id = NULL, action,
  params = NULL, ...)
}
\description{
This method allow the user to perform various analytics on
 varinats.
}
\details{
A method to query variant data
}
