% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaPanels-methods.R
\docType{methods}
\name{OpencgaPanels,Opencga-method}
\alias{OpencgaPanels,Opencga-method}
\alias{OpencgaCohorts}
\title{A method to query Opencga Panels}
\usage{
\S4method{OpencgaPanels}{Opencga}(object, id, action, params, ...)
}
\description{
This method allow the user to create, update and explore
cohorts data and metadta
}
