% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaTools-methods.R
\docType{methods}
\name{OpencgaTools,Opencga-method}
\alias{OpencgaTools,Opencga-method}
\alias{OpencgaCohorts}
\title{A method to query Opencga Tools}
\usage{
\S4method{OpencgaTools}{Opencga}(object, id, action, params, ...)
}
\description{
This method allow the user to create, update and explore
cohorts data and metadta
}
