% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpencgaVariableSet-methods.R
\docType{methods}
\name{OpencgaVariableSet,Opencga-method}
\alias{OpencgaVariableSet,Opencga-method}
\alias{OpencgaVariables}
\title{A method to query Opencga Variables}
\usage{
\S4method{OpencgaVariableSet}{Opencga}(object, id, action, params, ...)
}
\description{
This method allow the user to create, update and explore
vaiables data and metadta
}
